Move test to /tests and add lexer error
