+++ common/ast/Cargo.toml
@@ -8,19 +8,21 @@ edition = "2021"
 
 [lib]
 doctest = false
+test = false
 
 [dependencies] # In alphabetical order
 # Workspace dependencies
 common-exception = { path = "../exception" }
 common-functions = { path = "../functions" }
-logos = "0.12"
 
 # Github dependencies
 sqlparser = { git = "https://github.com/datafuse-extras/sqlparser-rs", rev = "c33837e" }
 
 # Crates.io dependencies
 async-trait = "0.1.52"
+logos = "0.12"
 once_cell = "1.9.0"
+thiserror = "1.0.30"
 
 [dev-dependencies]
 pretty_assertions = "1.0.0"
