+++ common/ast/src/error.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod ast;
-mod token;
-mod transformer;
-mod udfs;
+use thiserror::Error;
+
+// TODO: implement From<Result> for `common_exception::Result`.s
+pub type Result<T> = std::result::Result<T, Error>;
+
+/// Error is the error type for the common-ast crate.
+#[derive(Error, Debug, PartialEq, Eq)]
+pub enum Error {
+    #[error("unable to recognise the token from SQL: (rest {rest}, position {position})")]
+    UnrecognisedToken { rest: String, position: usize },
+}
