+++ common/ast/src/lib.rs
@@ -12,5 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod error;
 pub mod parser;
 pub mod udfs;
